added test case for input data added test case for models added main.py to recycle
