+++ recycle/main.py
@@ -1,0 +1,7 @@
+
+
+def get_input_data():
+    return {}
+
+def run_input_against_model(input_data, model):
+    return
